server/requirements.txt
uvicorn==0.25.0
fastapi[all]
pymongo
pydantic[email]
bcrypt
PyJWT
